server/requirements.txt
fastapi
uvicorn
openai
azure-cognitiveservices-speech